pyproject.toml
[project]
name = "rag-bios"
version = "0.1.0"
description = "Grounded RAG app for the BIOS technical challenge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.52.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.6",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.0",
    "faiss-cpu>=1.13.0",
    "tiktoken>=0.12.0",
    "python-dotenv>=1.1.1",
    "unstructured[all-docs]>=0.18.24",
    "openpyxl>=3.1.5",
]

[tool.uv]
package = false
